change to sql, new repository

+++ src/main/java/org/scholanova/mealdeliverapi/infrastructure/Menu/MenuRepository.java
@@ -3,5 +3,5 @@ package org.scholanova.mealdeliverapi.infrastructure.Menu;
 import org.scholanova.mealdeliverapi.domain.Menu;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MenuRepository extends CrudRepository<Menu, Integer> {
+public interface MenuRepository extends CrudRepository<Menu, Long> {
 }
