+++ src/main/java/org/scholanova/mealdeliverapi/infrastructure/Plat/repository/NourritureRepository.java
@@ -3,5 +3,5 @@ package org.scholanova.mealdeliverapi.infrastructure.Plat.repository;
 import org.scholanova.mealdeliverapi.domain.ItemNourriture.ItemNourriture;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NourritureRepository  extends CrudRepository<ItemNourriture, Integer> {
+public interface NourritureRepository  extends CrudRepository<ItemNourriture, Long> {
 }
